# src/immich_memories/web/openapi.json
{
  "components": {
    "schemas": {
...
      "Alternative": {
        "description": "Another picture of the shot's moment, eligible when the shot was chosen.",
        "properties": {
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "facts": {
            "title": "Facts",
            "type": "string"
          },
          "fate": {
            "title": "Fate",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "facts",
          "fate"
        ],
        "title": "Alternative",
        "type": "object"
      },
...
      "Cut": {
        "properties": {
          "content_budget_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Budget Seconds"
          },
          "content_seconds": {
            "title": "Content Seconds",
            "type": "number"
          },
          "film_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Film Seconds"
          },
          "model_polish": {
            "title": "Model Polish",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "shots": {
            "items": {
              "$ref": "#/components/schemas/CutShot"
            },
            "title": "Shots",
            "type": "array"
          },
          "thesis": {
            "title": "Thesis",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "thesis",
          "content_seconds",
          "content_budget_seconds",
          "film_seconds",
          "model_polish",
          "shots"
        ],
        "title": "Cut",
        "type": "object"
      },
...
      "CutShot": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/Alternative"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "chapter": {
            "title": "Chapter",
            "type": "string"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "moment": {
            "title": "Moment",
            "type": "string"
          },
          "motion": {
            "title": "Motion",
            "type": "boolean"
          },
          "new_day": {
            "title": "New Day",
            "type": "boolean"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "recorded_seconds": {
            "title": "Recorded Seconds",
            "type": "number"
          },
          "seconds": {
            "title": "Seconds",
            "type": "number"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionPath"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_interval": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Interval"
          },
          "start": {
            "title": "Start",
            "type": "number"
          },
          "story_key": {
            "title": "Story Key",
            "type": "string"
          },
          "story_title": {
            "title": "Story Title",
            "type": "string"
          },
          "taken": {
            "title": "Taken",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "position",
          "start",
          "seconds",
          "taken",
          "day",
          "new_day",
          "chapter",
          "story_key",
          "story_title",
          "moment",
          "reason",
          "motion",
          "recorded_seconds",
          "source_interval",
          "selection",
          "model",
          "alternatives"
        ],
        "title": "CutShot",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelDecision": {
        "description": "What the model polish recorded about a shot; absent when no model read the cut.",
        "properties": {
          "kept_reason": {
            "title": "Kept Reason",
            "type": "string"
          },
          "model_reason": {
            "title": "Model Reason",
            "type": "string"
          },
          "offered_count": {
            "title": "Offered Count",
            "type": "integer"
          },
          "proposed_asset_id": {
            "title": "Proposed Asset Id",
            "type": "string"
          },
          "replaced_asset_id": {
            "title": "Replaced Asset Id",
            "type": "string"
          },
          "replacement_outcome": {
            "title": "Replacement Outcome",
            "type": "string"
          },
          "seat": {
            "title": "Seat",
            "type": "string"
          }
        },
        "required": [
          "model_reason",
          "kept_reason",
          "proposed_asset_id",
          "offered_count",
          "replacement_outcome",
          "replaced_asset_id",
          "seat"
        ],
        "title": "ModelDecision",
        "type": "object"
      },
...
      "Revision": {
        "properties": {
          "content_seconds": {
            "title": "Content Seconds",
            "type": "number"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "removed": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Removed",
            "type": "array"
          },
          "segments": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "default": {},
            "title": "Segments",
            "type": "object"
          },
          "swaps": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Swaps",
            "type": "object"
          }
        },
        "required": [
          "number",
          "created_at",
          "content_seconds"
        ],
        "title": "Revision",
        "type": "object"
      },
      "RevisionEdits": {
        "description": "What the owner changed; every id must be in the cut, or a recorded sibling for a swap.",
        "properties": {
          "removed": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Removed",
            "type": "array"
          },
          "segments": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array"
            },
            "default": {},
            "title": "Segments",
            "type": "object"
          },
          "swaps": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Swaps",
            "type": "object"
          }
        },
        "title": "RevisionEdits",
        "type": "object"
      },
...
      "RunPage": {
        "properties": {
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "next_offset"
        ],
        "title": "RunPage",
        "type": "object"
      },
...
      "RunSummary": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "date_range_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range End"
          },
          "date_range_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Range Start"
          },
          "film": {
            "title": "Film",
            "type": "boolean"
          },
          "memory_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Type"
          },
          "preview_asset_ids": {
            "items": {
              "type": "string"
            },
            "title": "Preview Asset Ids",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "source",
          "created_at",
          "memory_type",
          "date_range_start",
          "date_range_end",
          "preview_asset_ids",
          "film"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "SelectionPath": {
        "description": "What the rules decided about a picture, in the words `runs why` prints.",
        "properties": {
          "facts": {
            "title": "Facts",
            "type": "string"
          },
          "kept_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kept At"
          },
          "left_out_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Out At"
          },
          "left_out_because": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Out Because"
          },
          "passed": {
            "items": {
              "type": "string"
            },
            "title": "Passed",
            "type": "array"
          }
        },
        "required": [
          "facts",
          "passed",
          "left_out_at",
          "left_out_because",
          "kept_at"
        ],
        "title": "SelectionPath",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Immich Memories",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "description": "Runs newest first, each with the first pictures its saved cut plays.",
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "completed",
                    "failed",
                    "running",
                    "cancelled",
                    "interrupted"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/cut": {
      "get": {
        "description": "The cut in the order it plays, each shot with every reason the run recorded for it.",
        "operationId": "read_cut_api_v1_runs__run_id__cut_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Cut",
        "tags": [
          "cut"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/revisions": {
      "get": {
        "description": "Every saved revision of this cut, oldest first.",
        "operationId": "list_revisions_api_v1_runs__run_id__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Revision"
                  },
                  "title": "Response List Revisions Api V1 Runs  Run Id  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Revisions",
        "tags": [
          "cut"
        ]
      },
      "post": {
        "description": "Keep the owner's edits as the next revision; 422 names the edit that would not render.",
        "operationId": "create_revision_api_v1_runs__run_id__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionEdits"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Revision"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Revision",
        "tags": [
          "cut"
        ]
      }
    },
...
  }
}